# slightly optimize tflite Concatenate operation