@stnk20 2019-09-26
> Hi, thanks for this change! It looks good. If you want, there appears to be a similar issue with ConcatenationWithScaling, so there might be a similar gain there.

I agree, but I don't understand the way for testing and checking performance.
Is there a simple example for this?